Added floor actuators, 0 level ready, 1 current

+++ DungeonMaster/DungeonMaster/Descriptors/DescriptorBase.cs
@@ -4,7 +4,7 @@ namespace DungeonMasterParser.Descriptors
     {
         public int Identifer { get; set; } = -1;
         public string Name { get; set; }
-        public float Weight { get; set; }
+        public int Weight { get; set; }
         public string TexturePath { get; set; }
 
         public override string ToString() => $"{Identifer} {Name}";
